Lv 2 Platform
Includes a lv 2 platform. Got rid of camera lag because it didn't work with the falling, that code needs to be changed to accustom the falling elevation. Question difficulty to successfully pass the falling part. Tested with lights on and was possible, the questionable part is how possible it is with lights off. Haven't had any successful with lights turned off.